MB-7297 Use size_t for python wrapping for 32-bit support
The couchstore/couch_db.h uses size_t for several struct fields, and
size_t can be either 32-bits or 64-bits, depending on platform.
So, instead of c_ulonglong, use c_size_t for those fields.

Here, we assume sizeof(off_t) == sizeof(size_t). Not a 100%
bulletproof assumption, but it seems to work on tests.

Additionally, use uint64_t for the couchstore_open_db() flags
parameter rather than the platform-specific int.

Reviewed-on: http://review.couchbase.org/22969
Tested-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>

+++ python/couchstore.py
@@ -65,7 +65,7 @@ def _toString (key):
 ### INTERNAL STRUCTS:
 
 class SizedBuf (Structure):
-    _fields_ = [("buf", POINTER(c_char)), ("size", c_ulonglong)]
+    _fields_ = [("buf", POINTER(c_char)), ("size", c_size_t)]
 
     def __init__(self, string):
         if string != None:
@@ -90,7 +90,7 @@ class DocInfoStruct (Structure):
                 ("deleted", c_int),
                 ("content_meta", c_ubyte),
                 ("bp", c_ulonglong),
-                ("size", c_ulonglong) ]
+                ("size", c_size_t) ]
 
 class LocalDocStruct (Structure):
     _fields_ = [("id", SizedBuf),
@@ -103,7 +103,7 @@ class DbInfoStruct (Structure):
                 ("doc_count", c_ulonglong),
                 ("deleted_count", c_ulonglong),
                 ("space_used", c_ulonglong),
-                ("header_position", c_ulonglong) ]
+                ("header_position", c_size_t) ]
 
 
 ### DOCUMENT INFO CLASS:
@@ -223,7 +223,7 @@ class CouchStore (object):
             flags = 0
 
         db = c_void_p()
-        _check(_lib.couchstore_open_db(path, flags, byref(db)))
+        _check(_lib.couchstore_open_db(path, c_uint64(flags), byref(db)))
         self._as_parameter_ = db
         self.path = path
